% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptconcept2_sFFLHD_R6_desfuncs.R
\name{actual_des_func_grad_norm2_mean_gramacy2Dexp3hole}
\alias{actual_des_func_grad_norm2_mean_gramacy2Dexp3hole}
\title{actual des func}
\usage{
actual_des_func_grad_norm2_mean_gramacy2Dexp3hole(XX, mod)
}
\arguments{
\item{XX}{design matrix}

\item{mod}{mod}
}
\value{
values
}
\description{
actual des func
}
