% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LHS.R
\name{phi_p}
\alias{phi_p}
\title{phi_p}
\usage{
phi_p(X, p = 50, t = 1)
}
\arguments{
\item{X}{design matrix}

\item{p}{Power}

\item{t}{distance metric, don't change}
}
\value{
design phi_p value
}
\description{
phi_p
}
