% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptconcept2_sFFLHD_R6_desfuncs.R
\name{test_des_func_grad_norm2_mean}
\alias{test_des_func_grad_norm2_mean}
\title{Test des func}
\usage{
test_des_func_grad_norm2_mean(func, actual, d, n = 1000)
}
\arguments{
\item{func}{function}

\item{actual}{actual values}

\item{d}{dimension}

\item{n}{number of points}
}
\value{
value
}
\description{
Test des func
}
