% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/adaptconcept2_sFFLHD_R6_desfuncs.R
\name{werror_func_relmax}
\alias{werror_func_relmax}
\title{werror_func_relmax}
\usage{
werror_func_relmax(mod, XX, alpha = 1000, split_speed = T)
}
\arguments{
\item{mod}{model}

\item{XX}{XX}

\item{alpha}{alpha}

\item{split_speed}{T/F}
}
\value{
werror
}
\description{
werror_func_relmax
}
\examples{
# pass
}
