% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compare_adaptconceptR6.R
\docType{class}
\name{compare.adaptR6}
\alias{compare.adaptR6}
\title{Class providing object with methods for adapt.concept2.sFFLHD.R6}
\format{\code{\link{R6Class}} object.}
\value{
Object of \code{\link{R6Class}} with methods for running an
adaptive experiment.
}
\description{
Class providing object with methods for adapt.concept2.sFFLHD.R6

Class providing object with methods for adapt.concept2.sFFLHD.R6
}
\section{Methods}{

\describe{
  \item{Documentation}{For full documentation of each method go to
  https://github.com/CollinErickson/DOE-Code}
  \item{\code{new(X, Z, corr="Gauss", verbose=0, separable=T,
  useC=F,useGrad=T,
         parallel=T, nug.est=T, ...)}}{This method is used to create object
         of this class with \code{X} and \code{Z} as the data.}

  \item{\code{update(Xnew=NULL, Znew=NULL, Xall=NULL, Zall=NULL,
restarts = 5,
param_update = T, nug.update = self$nug.est)}}{This method updates the
model, adding new data if given, then running optimization again.}
  }
}

\examples{
ca1 <- compare.adaptR6$new(func=TestFunctions::gaussian1, D=2, L=3,
                           batches=2, reps=2,
                           n0=6, obj="desirability",
                           selection_method=c('max_des', 'SMED'),
                           des_func=c('des_func_relmax', 'des_func_relmax')
                           )$run_all()$plot()

}
\keyword{Gaussian}
\keyword{adaptive,}
\keyword{data,}
\keyword{experiments,}
\keyword{process,}
\keyword{regression}
\keyword{sequential,}
\keyword{simulation,}
\section{Public fields}{
\if{html}{\out{<div class="r6-fields">}}
\describe{
\item{\code{L}}{batch size}

\item{\code{batches}}{Number of batches to run each for}

\item{\code{reps}}{Number of replicates of each}

\item{\code{forces}}{Can force values}

\item{\code{force_vals}}{Can force values}

\item{\code{b}}{batch size}

\item{\code{func}}{Actual function to get experiment values from}

\item{\code{D}}{Dimension of data}

\item{\code{package}}{Which GP package to use in IGP}

\item{\code{design}}{A string saying which design object should be used.}

\item{\code{obj}}{A string saying what the objective is.}

\item{\code{n0}}{The initial number of points to be selected.}

\item{\code{force_old}}{A number saying how often the oldest candidate points
should be forced into the design.}

\item{\code{force_pvar}}{A number saying how often the points with the highest
predictive variance should be forced into the design.}

\item{\code{des_func}}{The desirability function.}

\item{\code{alpha_des}}{The alpha constant for the weight function.}

\item{\code{actual_des_func}}{The true des func used to evaluate the model,
not known in practice}

\item{\code{weight_const}}{The weight constant in the weight function, usually 1.}

\item{\code{selection_method}}{What the selection method should be.}

\item{\code{parallel}}{Should new values be calculated in parallel?}

\item{\code{seed_start}}{A seed to be set for each run}

\item{\code{func_string}}{The name of the function}

\item{\code{save_output}}{Should the output be saved regularly?}

\item{\code{stage1batches}}{Number of batches to be run in "stage 1,"
aka nonadaptively, chosen to be space filling}

\item{\code{folder_name}}{Where should the files be saved?}

\item{\code{folder_path}}{What is the path for the folder?}

\item{\code{design_seed_start}}{Random seed for the designs}

\item{\code{outdf}}{A data frame with all the output}

\item{\code{outrawdf}}{A data frame with all the raw output}

\item{\code{plotdf}}{A data frame with data to be plotted}

\item{\code{enddf}}{A data frame with the output at the end of each simulation}
}
\if{html}{\out{</div>}}
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{compare.adaptR6$new()}}
\item \href{#method-set_folder_name}{\code{compare.adaptR6$set_folder_name()}}
\item \href{#method-create_output_folder}{\code{compare.adaptR6$create_output_folder()}}
\item \href{#method-run_all}{\code{compare.adaptR6$run_all()}}
\item \href{#method-run_one}{\code{compare.adaptR6$run_one()}}
\item \href{#method-add_result_of_one}{\code{compare.adaptR6$add_result_of_one()}}
\item \href{#method-postprocess_outdf}{\code{compare.adaptR6$postprocess_outdf()}}
\item \href{#method-plot_MSE_over_batch}{\code{compare.adaptR6$plot_MSE_over_batch()}}
\item \href{#method-plot_AWE_over_batch}{\code{compare.adaptR6$plot_AWE_over_batch()}}
\item \href{#method-plot_AWE_over_group}{\code{compare.adaptR6$plot_AWE_over_group()}}
\item \href{#method-plot_MSE_PVar}{\code{compare.adaptR6$plot_MSE_PVar()}}
\item \href{#method-plot_RMSE_PRMSE}{\code{compare.adaptR6$plot_RMSE_PRMSE()}}
\item \href{#method-plot}{\code{compare.adaptR6$plot()}}
\item \href{#method-plot_run_times}{\code{compare.adaptR6$plot_run_times()}}
\item \href{#method-save_self}{\code{compare.adaptR6$save_self()}}
\item \href{#method-delete_save_folder_if_empty}{\code{compare.adaptR6$delete_save_folder_if_empty()}}
\item \href{#method-recover_parallel_temp_save}{\code{compare.adaptR6$recover_parallel_temp_save()}}
\item \href{#method-parallel_efficiency}{\code{compare.adaptR6$parallel_efficiency()}}
\item \href{#method-clone}{\code{compare.adaptR6$clone()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\subsection{Method \code{new()}}{
Initialize the R6 object
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$new(
  func,
  D,
  L,
  b = NULL,
  batches = 10,
  reps = 5,
  obj = c("nonadapt", "grad"),
  force_old = c(0),
  force_pvar = c(0),
  n0 = 0,
  stage1batches = 0,
  save_output = F,
  func_string = NULL,
  seed_start = as.numeric(Sys.time()),
  design_seed_start = as.numeric(Sys.time()),
  package = "laGP",
  selection_method = "SMED",
  design = "sFFLHD",
  des_func = NA,
  alpha_des = 1,
  weight_const = 0,
  error_power = 1,
  actual_des_func = NULL,
  pass_list = list(),
  folder_name,
  parallel = FALSE,
  parallel_cores = "detect"
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{func}}{The true function. Should take a matrix as input with
D columns, each row is an X point. Or just each point as a vector,
depends on func_run_together}

\item{\code{D}}{Number of input dimensions. Should all be 0 to 1.}

\item{\code{L}}{Batch size}

\item{\code{b}}{Batch size to add each iteration}

\item{\code{batches}}{Number of batches to run for}

\item{\code{reps}}{Number of replicates of each combination}

\item{\code{obj}}{Objective type}

\item{\code{force_old}}{Proportion of points to be force added
from Xopts because they are old.}

\item{\code{force_pvar}}{Proportion of points to be force added
from Xopts because they have highest predictive variance}

\item{\code{n0}}{Number of points to start experiment with}

\item{\code{stage1batches}}{Number of batches to be run in "stage 1,"
aka nonadaptively, chosen to be space filling}

\item{\code{save_output}}{Should the output be saved regularly?}

\item{\code{func_string}}{The name of the function}

\item{\code{seed_start}}{A seed to be set for each run}

\item{\code{design_seed_start}}{A seed to be set for the design points}

\item{\code{package}}{Gaussian process model package to use}

\item{\code{selection_method}}{How should points be selected?}

\item{\code{design}}{The design to take candidate points from.}

\item{\code{des_func}}{The desirability function}

\item{\code{alpha_des}}{Alpha value for the desirability function}

\item{\code{weight_const}}{What is the weight constant?}

\item{\code{error_power}}{Which error power should be saved? Either 0, 1, or
maybe c(0,1)?}

\item{\code{actual_des_func}}{The actual des func}

\item{\code{pass_list}}{List of things to pass to adapt concept for each.}

\item{\code{folder_name}}{Where should the files be saved?}

\item{\code{parallel}}{Should points be evaluated in parallel}

\item{\code{parallel_cores}}{Number of parallel cores to be used.}

\item{\code{useSMEDtheta}}{Should theta be used when SMED is used?
Theta is the correlation parameter in each dimension. Helps
space things properly}

\item{\code{take_until_maxpvar_below}}{If given, nonadaptive batches will
be taken instead of adaptive until max pvar is below this value.}

\item{\code{new_batches_per_batch}}{Each time a batch is added,
how many batches of points should be added as candidates}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_folder_name"></a>}}
\subsection{Method \code{set_folder_name()}}{
Set the folder name where this will be saved
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$set_folder_name(folder_name, add_timestamp = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{folder_name}}{Name of the folder to use}

\item{\code{add_timestamp}}{Should the timestamp be included in the folder name?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_output_folder"></a>}}
\subsection{Method \code{create_output_folder()}}{
Create the folder to save the output
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$create_output_folder(add_timestamp = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{add_timestamp}}{Not used.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_all"></a>}}
\subsection{Method \code{run_all()}}{
Run all remaining simulations
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$run_all(
  redo = FALSE,
  noplot = FALSE,
  save_every = FALSE,
  run_order,
  parallel = self$parallel,
  parallel_temp_save = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{redo}}{Should already completed ones be redone?}

\item{\code{noplot}}{Should plots not be made?}

\item{\code{save_every}}{Should the object be saved after each simulation,
or just at the end?}

\item{\code{run_order}}{In what order should they be run?}

\item{\code{parallel}}{Should it be run in parallel?}

\item{\code{parallel_temp_save}}{Should temporary files be saved by each
core when running in parallel? Useful if it crashes in the middle,
can just recover from the files afterward.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-run_one"></a>}}
\subsection{Method \code{run_one()}}{
Run a single simulation
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$run_one(
  irow = NULL,
  save_output = self$save_output,
  noplot = FALSE,
  is_parallel = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{irow}}{The simulation number to be run}

\item{\code{save_output}}{Should the output be saved?}

\item{\code{noplot}}{Should the plots not be made?}

\item{\code{is_parallel}}{Is it running in a parallel environment?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_result_of_one"></a>}}
\subsection{Method \code{add_result_of_one()}}{
Add result of one simulation run to the results
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$add_result_of_one(irow, newdf1, save_output = self$save_output)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{irow}}{The simulation number that was run}

\item{\code{newdf1}}{The output data frame from that run}

\item{\code{save_output}}{Should the output be saved?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-postprocess_outdf"></a>}}
\subsection{Method \code{postprocess_outdf()}}{
Postprocess the output in outdf
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$postprocess_outdf(save_output = self$save_output)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{save_output}}{Should the output be saved?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_MSE_over_batch"></a>}}
\subsection{Method \code{plot_MSE_over_batch()}}{
Plot MSE by batch
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$plot_MSE_over_batch(
  save_output = self$save_output,
  legend_labels = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{save_output}}{Should the plot be saved?}

\item{\code{legend_labels}}{Override the legend labels}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_AWE_over_batch"></a>}}
\subsection{Method \code{plot_AWE_over_batch()}}{
Ploted average weighted integrated weighted error
by batch.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$plot_AWE_over_batch(save_output = self$save_output)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{save_output}}{Should the plot be saved?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_AWE_over_group"></a>}}
\subsection{Method \code{plot_AWE_over_group()}}{
Plot actual integrated weighted error by group
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$plot_AWE_over_group(
  save_output = self$save_output,
  boxpl = TRUE,
  logy = TRUE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{save_output}}{Should the plot be saved?}

\item{\code{boxpl}}{Should it be a box plot?}

\item{\code{logy}}{Should the y axis be on the log scale?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_MSE_PVar"></a>}}
\subsection{Method \code{plot_MSE_PVar()}}{
Plot Predicted variance vs mean squared error
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$plot_MSE_PVar(save_output = self$save_output)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{save_output}}{Should the plot be saved?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_RMSE_PRMSE"></a>}}
\subsection{Method \code{plot_RMSE_PRMSE()}}{
Plot Predicted RMSE vs RMSE
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$plot_RMSE_PRMSE(save_output = self$save_output)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{save_output}}{Should the plot be saved?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot"></a>}}
\subsection{Method \code{plot()}}{
Make plots
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$plot(save_output = self$save_output)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{save_output}}{Should the plots be saved?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-plot_run_times"></a>}}
\subsection{Method \code{plot_run_times()}}{
Plot the time it took to run each one.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$plot_run_times()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-save_self"></a>}}
\subsection{Method \code{save_self()}}{
Save this object to file.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$save_self(object_name = "object")}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{object_name}}{Name of object to save in file.
By default it is "object".}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete_save_folder_if_empty"></a>}}
\subsection{Method \code{delete_save_folder_if_empty()}}{
Deletes the save folder if empty.
Prevents empty folders from being left behind.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$delete_save_folder_if_empty()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-recover_parallel_temp_save"></a>}}
\subsection{Method \code{recover_parallel_temp_save()}}{
Recover temporary files that were saved while
running in parallel if it crashed previously while running.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$recover_parallel_temp_save(save_if_any_recovered = TRUE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{save_if_any_recovered}}{Should the object be saved immediately
if it was able to recover files?}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-parallel_efficiency"></a>}}
\subsection{Method \code{parallel_efficiency()}}{
Calculate the parallel efficiency
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$parallel_efficiency()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-clone"></a>}}
\subsection{Method \code{clone()}}{
The objects of this class are cloneable with this method.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{compare.adaptR6$clone(deep = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{deep}}{Whether to make a deep clone.}
}
\if{html}{\out{</div>}}
}
}
}
